# src/adhara/interop-service/crosschainInterop/api/admin/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Administration API",
    "description": "Crosschain interop service Administration API defines endpoints for maintaining the Corda Notaries, Corda Participants, Corda Registered Functions, Interop Authentication Parameters, Interop Participants, and Validators resources."
  },
  "host": "localhost:3031",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/{networkId}/cordaParticipants": {
      "get": {
        "tags": [
          "CordaParticipants"
        ],
        "summary": "Fetched if is a Corda participant.",
        "description": "",
        "operationId": "getCordaParticipant",
        "parameters": [
          {
            "name": "networkId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "remoteNetworkId",
            "in": "query",
            "required": true,
            "type": "number"
          },
          {
            "name": "publicKey",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched if is a Corda participant.",
            "schema": {
              "type": "object",
              "properties": {
                "isParticipant": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "An error occurred."
          }
        }
      },
      "post": {
        "tags": [
          "CordaParticipants"
        ],
        "summary": "Create a corda participant.",
        "description": "",
        "operationId": "postCordaParticipant",
        "parameters": [
          {
            "name": "networkId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "cordaParticipant",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CordaParticipant"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created corda participant.",
            "schema": {
              "$ref": "#/definitions/CordaParticipant"
            }
          },
          "500": {
            "description": "An error occurred."
          }
        }
      },
      "delete": {
        "tags": [
          "CordaParticipants"
        ],
        "summary": "Remove a corda participant.",
        "description": "",
        "operationId": "deleteCordaParticipant",
        "parameters": [
          {
            "name": "networkId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "remoteNetworkId",
            "in": "query",
            "required": true,
            "type": "number"
          },
          {
            "name": "publicKey",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Removed corda participant.",
            "schema": {
              "$ref": "#/definitions/CordaParticipant"
            }
          },
          "500": {
            "description": "An error occurred."
          }
        }
      }
    },
...
    "/{networkId}/interopParticipants": {
      "get": {
        "tags": [
          "InteropParticipants"
        ],
        "summary": "Fetch the interop participant.",
        "description": "",
        "operationId": "getLocalAccountId",
        "parameters": [
          {
            "name": "networkId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "remoteAccountId",
            "in": "query",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched the interop participant.",
            "schema": {
              "$ref": "#/definitions/InteropParticipant"
            }
          },
          "500": {
            "description": "An error occurred."
          }
        }
      },
      "post": {
        "tags": [
          "InteropParticipants"
        ],
        "summary": "Create interop participant.",
        "description": "",
        "operationId": "postInteropParticipants",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "networkId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "interopParticipant",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/InteropParticipant"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created interop participant.",
            "schema": {
              "$ref": "#/definitions/InteropParticipant"
            }
          },
          "500": {
            "description": "An error occurred."
          }
        }
      },
      "delete": {
        "tags": [
          "InteropParticipants"
        ],
        "summary": "Remove interop participant.",
        "description": "",
        "operationId": "deleteInteropParticipant",
        "parameters": [
          {
            "name": "networkId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "remoteAccountId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "interopParticipant",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/InteropParticipant"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed interop participant.",
            "schema": {
              "$ref": "#/definitions/InteropParticipant"
            }
          },
          "500": {
            "description": "An error occurred."
          }
        }
      }
    },
    "/{networkId}/interopAuthParams": {
      "get": {
        "tags": [
          "InteropAuthParams"
        ],
        "summary": "Fetch if is interop authentication parameters.",
        "description": "",
        "operationId": "getInteropAuthParams",
        "parameters": [
          {
            "name": "networkId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "remoteNetworkId",
            "in": "query",
            "required": true,
            "type": "number"
          },
          {
            "name": "remoteContractAddress",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched if is interop authentication parameters.",
            "schema": {
              "type": "object",
              "properties": {
                "isAuthParam": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "An error occurred."
          }
        }
      },
      "post": {
        "tags": [
          "InteropAuthParams"
        ],
        "summary": "Create interop authentication parameters.",
        "description": "",
        "operationId": "postInteropAuthParams",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "networkId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "interopAuthParam",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/InteropAuthParam"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created interop authentication parameters.",
            "schema": {
              "$ref": "#/definitions/InteropAuthParam"
            }
          },
          "500": {
            "description": "An error occurred."
          }
        }
      },
      "delete": {
        "tags": [
          "InteropAuthParams"
        ],
        "summary": "Remove interop authentication parameters.",
        "description": "",
        "operationId": "deleteInteropAuthParams",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "networkId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "interopAuthParam",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/InteropAuthParam"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed interop authentication parameters.",
            "schema": {
              "$ref": "#/definitions/InteropAuthParam"
            }
          },
          "500": {
            "description": "An error occurred."
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "InteropAuthParam": {
      "type": "object",
      "properties": {
        "remoteNetworkId": {
          "type": "string",
          "example": "string"
        },
        "remoteContractAddress": {
          "type": "string",
          "example": "string"
        }
      },
      "required": [
        "remoteNetworkId",
        "remoteContractAddress"
      ]
    },
    "InteropParticipant": {
      "type": "object",
      "properties": {
        "localAccountId": {
          "type": "string",
          "example": "string"
        },
        "remoteAccountId": {
          "type": "string",
          "example": "string"
        }
      },
      "required": [
        "localAccountId",
        "remoteAccountId"
      ]
    },
...
    "CordaParticipant": {
      "type": "object",
      "properties": {
        "remoteNetworkId": {
          "type": "number",
          "example": 0
        },
        "publicKey": {
          "type": "string",
          "example": "string"
        }
      },
      "required": [
        "remoteNetworkId",
        "publicKey"
      ]
    },
...
  }
}